Minor script improvements
Add --bitrate switch to canlogger and canplayer
Execute canlogger in single thread for better exception handling
Move MessageSync class to can.io.player module
